Build using Sphinx, update logo, and refresh design
A built version of this PR is available at: <https://toltec-dev.org/pr-sphinx/>

Build system change
-------------------

This PR changes the build system from docutils to Sphinx. This will make it easier to structure the website in several pages in the future, as I think it would be useful to have other pages with more detailed user instructions. This also makes customizing the HTML output and build process easier and cleaner.

* The overall site structure is defined in [`src/sitemap.rst`](src/sitemap.rst). This file is used to generate the sidebar.
* The global site layout is defined in [`src/_themes/toltec/layout.html`](src/_themes/toltec/layout.html). In that folder, the stylesheet and global images can be found.
* Toltec-specific Sphinx extensions are defined under [`src/_ext`](src/_ext).

To build the website, run `make prod`. To start a development server with live-reloading, run `make dev`.

Design changes
--------------

I added a sidebar that will prove useful in the future if we add other pages to the website. The stylesheet now respects the user’s light or dark theme preference. The logo was updated to the one from toltec-dev/organization#30.