workflows: Allow setting the SSH port for the sync action
I recently migrated the Toltec website to a different server in which
the SSH server listens on a non-standard port. This PR changes our CI
workflow to allow specifying this port when syncing stable/testing
updates to the website. The port is set through a secret called SSH_PORT
which I have already registered.